+++ crates/sui-adapter/src/error.rs
@@ -11,10 +11,8 @@ use move_core_types::{
     vm_status::{StatusCode, StatusType},
 };
 use move_vm_runtime::move_vm::MoveVM;
-use sui_types::{
-    error::{ExecutionError, SuiError},
-    messages::{ExecutionFailureStatus, MoveLocation, MoveLocationOpt},
-};
+use sui_types::error::{ExecutionError, SuiError};
+use sui_types::execution_status::{ExecutionFailureStatus, MoveLocation, MoveLocationOpt};
 
 pub(crate) fn convert_vm_error<S: MoveResolver<Err = SuiError>>(
     error: VMError,
