+++ crates/sui-core/src/generate_format.rs
@@ -13,6 +13,10 @@ use rand::SeedableRng;
 use serde_reflection::{Registry, Result, Samples, Tracer, TracerConfig};
 use shared_crypto::intent::{Intent, IntentMessage, PersonalMessage};
 use std::{fs::File, io::Write};
+use sui_types::execution_status::{
+    CommandArgumentError, ExecutionFailureStatus, ExecutionStatus, PackageUpgradeError,
+    TypeArgumentError,
+};
 use sui_types::{
     base_types::MoveObjectType_,
     crypto::Signer,
@@ -30,10 +34,7 @@ use sui_types::{
         get_key_pair, get_key_pair_from_rng, AccountKeyPair, AuthorityKeyPair,
         AuthorityPublicKeyBytes, AuthoritySignature, KeypairTraits, Signature, SuiKeyPair,
     },
-    messages::{
-        Argument, CallArg, Command, CommandArgumentError, ExecutionFailureStatus, ExecutionStatus,
-        ObjectArg, ObjectInfoRequestKind, PackageUpgradeError, TransactionKind, TypeArgumentError,
-    },
+    messages::{Argument, CallArg, Command, ObjectArg, ObjectInfoRequestKind, TransactionKind},
     multisig::{MultiSig, MultiSigPublicKey},
     object::{Data, Owner},
     signature::GenericSignature,
