+++ crates/sui-core/src/unit_tests/move_package_tests.rs
@@ -10,12 +10,12 @@ use sui_types::{
     base_types::ObjectID,
     digests::TransactionDigest,
     error::{ExecutionErrorKind, SuiError},
-    messages::PackageUpgradeError,
     move_package::{MovePackage, TypeOrigin, UpgradeInfo},
     object::{Data, Object, OBJECT_START_VERSION},
 };
 
 use std::{collections::BTreeMap, path::PathBuf};
+use sui_types::execution_status::PackageUpgradeError;
 
 macro_rules! type_origin_table {
     {} => { Vec::new() };
