+++ crates/sui-json-rpc-types/src/sui_transaction.rs
@@ -24,12 +24,13 @@ use sui_types::base_types::{
 };
 use sui_types::digests::{ObjectDigest, TransactionEventsDigest};
 use sui_types::error::{ExecutionError, SuiError};
+use sui_types::execution_status::ExecutionStatus;
 use sui_types::gas::GasCostSummary;
 use sui_types::messages::{
-    Argument, CallArg, Command, ExecuteTransactionRequestType, ExecutionStatus, GenesisObject,
-    InputObjectKind, ObjectArg, ProgrammableMoveCall, ProgrammableTransaction, SenderSignedData,
-    TransactionData, TransactionDataAPI, TransactionEffects, TransactionEffectsAPI,
-    TransactionEvents, TransactionKind, VersionedProtocolMessage,
+    Argument, CallArg, Command, ExecuteTransactionRequestType, GenesisObject, InputObjectKind,
+    ObjectArg, ProgrammableMoveCall, ProgrammableTransaction, SenderSignedData, TransactionData,
+    TransactionDataAPI, TransactionEffects, TransactionEffectsAPI, TransactionEvents,
+    TransactionKind, VersionedProtocolMessage,
 };
 use sui_types::messages_checkpoint::CheckpointSequenceNumber;
 use sui_types::move_package::disassemble_modules;
