+++ crates/sui-json-rpc/src/balance_changes.rs
@@ -16,8 +16,9 @@ use sui_types::base_types::{ObjectID, ObjectRef, SequenceNumber};
 use sui_types::coin::Coin;
 use sui_types::digests::ObjectDigest;
 use sui_types::error::SuiError;
+use sui_types::execution_status::ExecutionStatus;
 use sui_types::gas_coin::GAS;
-use sui_types::messages::{ExecutionStatus, TransactionEffects};
+use sui_types::messages::TransactionEffects;
 use sui_types::messages::{InputObjectKind, TransactionEffectsAPI};
 use sui_types::object::{Object, Owner};
 use sui_types::storage::WriteKind;
