+++ crates/sui-transactional-test-runner/src/test_adapter.rs
@@ -48,15 +48,13 @@ use sui_framework::BuiltInFramework;
 use sui_framework::DEFAULT_FRAMEWORK_PATH;
 use sui_protocol_config::ProtocolConfig;
 use sui_types::accumulator::Accumulator;
+use sui_types::execution_status::ExecutionStatus;
 use sui_types::MOVE_STDLIB_OBJECT_ID;
 use sui_types::{
     base_types::{ObjectID, ObjectRef, SuiAddress, TransactionDigest, SUI_ADDRESS_LENGTH},
     crypto::{get_key_pair_from_rng, AccountKeyPair},
     event::Event,
-    messages::{
-        ExecutionStatus, TransactionData, TransactionDataAPI, TransactionEffectsAPI,
-        VerifiedTransaction,
-    },
+    messages::{TransactionData, TransactionDataAPI, TransactionEffectsAPI, VerifiedTransaction},
     object::{self, Object, ObjectFormatOptions},
     object::{MoveObject, Owner},
     MOVE_STDLIB_ADDRESS, SUI_CLOCK_OBJECT_ID, SUI_CLOCK_OBJECT_SHARED_VERSION,
