+++ crates/sui-types/src/temporary_store.rs
@@ -15,6 +15,7 @@ use sui_protocol_config::ProtocolConfig;
 use tracing::trace;
 
 use crate::committee::EpochId;
+use crate::execution_status::ExecutionStatus;
 use crate::messages::TransactionEvents;
 use crate::storage::ObjectStore;
 use crate::sui_system_state::{
@@ -28,7 +29,7 @@ use crate::{
     event::Event,
     fp_bail, gas,
     gas::{GasCostSummary, SuiGasStatus, SuiGasStatusAPI},
-    messages::{ExecutionStatus, InputObjects, TransactionEffects},
+    messages::{InputObjects, TransactionEffects},
     object::Owner,
     object::{Data, Object},
     storage::{
