+++ crates/transaction-fuzzer/src/executor.rs
@@ -13,9 +13,7 @@ use sui_types::object::Owner;
 use sui_types::utils::to_sender_signed_transaction;
 use sui_types::{
     error::SuiError,
-    messages::{
-        ExecutionFailureStatus, ExecutionStatus, TransactionEffectsAPI, VerifiedTransaction,
-    },
+    messages::{TransactionEffectsAPI, VerifiedTransaction},
     object::Object,
 };
 use tokio::runtime::Runtime;
@@ -24,6 +22,7 @@ use crate::account_universe::{AccountCurrent, INITIAL_BALANCE};
 
 use std::path::PathBuf;
 use sui_move_build::BuildConfig;
+use sui_types::execution_status::{ExecutionFailureStatus, ExecutionStatus};
 
 pub type ExecutionResult = Result<ExecutionStatus, SuiError>;
